requirements.txt
discord.py>=2.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
